fix: move audit folder
